pyproject.toml
[project]
name = "marketplace-backend"
version = "0.1.0"
description = "Marketplace Prototype API (FastAPI + SQLAlchemy 2.0 + MinIO)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "sqlalchemy>=2.0.28",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "aiobotocore>=2.12.2", # Асинхронный клиент для работы с S3/MinIO
    "python-multipart>=0.0.9", # Обязателен для эндпоинта загрузки изображений (UploadFile)
    "pyrefly>=0.55.0",
    "pyjwt>=2.11.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "boto3>=1.42.61",
]

[tool.ruff]
# Указываем целевую версию, чтобы Ruff использовал фичи Python 3.12
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
"E",             # Стандартные ошибки PEP 8 (pycodestyle)
"W",             # Предупреждения PEP 8 (pycodestyle)
"F",             # Логические ошибки (pyflakes)
"I",             # Сортировка импортов (isort)
"UP",            # pyupgrade — АВТОМАТИЧЕСКИ заменяет typing.Optional на `| None` и typing.List на `list`
"B",             # Поиск частых багов (flake8-bugbear)
"C4",            # Оптимизация генераторов и словарей (flake8-comprehensions)
"SIM",           # Упрощение кода (flake8-simplify)
"ASYNC",         # flake8-async — КРИТИЧЕСКИ ВАЖНО для нас: ловит ошибки работы с asyncio
]
ignore = [
"B008",          # Do not perform function call in argument defaults (standard for FastAPI Depends)
]

# Строго запрещаем использовать устаревшие аннотации из модуля typing
[tool.ruff.lint.flake8-tidy-imports.banned-api]
"typing.Optional".msg = "В Python 3.12 используй `T | None`."
"typing.Union".msg = "В Python 3.12 используй `T | U`."
"typing.List".msg = "В Python 3.12 используй встроенный `list`."
"typing.Dict".msg = "В Python 3.12 используй встроенный `dict`."
"typing.Tuple".msg = "В Python 3.12 используй встроенный `tuple`."

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
